+++ bootique-bom-tests-all/src/main/java/io/bootique/bom/jersey/JerseyAppFeature.java
@@ -19,8 +19,8 @@
 
 package io.bootique.bom.jersey;
 
-import javax.ws.rs.core.Feature;
-import javax.ws.rs.core.FeatureContext;
+import jakarta.ws.rs.core.Feature;
+import jakarta.ws.rs.core.FeatureContext;
 
 public class JerseyAppFeature implements Feature {