+++ bootique-bom-tests-all/src/main/java/io/bootique/bom/jersey/r2/Resource2.java
@@ -20,12 +20,12 @@
 package io.bootique.bom.jersey.r2;
 
 import io.bootique.annotation.Args;
-
 import jakarta.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+
 import java.util.Arrays;
 
 import static java.util.stream.Collectors.joining;