+++ bootique-bom-tests-all/src/main/java/io/bootique/bom/jetty/BomFilter.java
@@ -19,15 +19,11 @@
 
 package io.bootique.bom.jetty;
 
+import jakarta.servlet.*;
+import jakarta.servlet.annotation.WebFilter;
+
 import java.io.IOException;
 
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.annotation.WebFilter;
 
 @WebFilter(filterName = "bom", urlPatterns = "/*")
 public class BomFilter implements Filter {