+++ bootique-bom-tests-all/src/main/java/io/bootique/bom/jetty/BomServlet.java
@@ -19,13 +19,12 @@
 
 package io.bootique.bom.jetty;
 
-import java.io.IOException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
 @WebServlet(name = "bom", urlPatterns = "/*")
 public class BomServlet extends HttpServlet {
@@ -33,8 +32,7 @@ public class BomServlet extends HttpServlet {
 	private static final long serialVersionUID = 3100731292623511396L;
 
 	@Override
-	protected void doGet(HttpServletRequest request, HttpServletResponse response)
-			throws ServletException, IOException {
+	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		response.getWriter().println("bom_servlet_query_string: " + request.getQueryString());
 	}
 }