+++ bootique-bom-tests-all/src/main/java/io/bootique/bom/mvc/r1/MustacheResource.java
@@ -19,8 +19,9 @@
 
 package io.bootique.bom.mvc.r1;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
+
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
 
 @Path("/mustache")
 public class MustacheResource {