+++ bootique-bom-tests-all/src/test/java/io/bootique/bom/jersey/JerseyAppJunit4IT.java
@@ -26,13 +26,11 @@ import io.bootique.jersey.JerseyModule;
 import io.bootique.jetty.JettyModule;
 import io.bootique.test.junit.BQTestFactory;
 import io.bootique.test.junit.TestIO;
+import jakarta.ws.rs.client.ClientBuilder;
+import jakarta.ws.rs.client.WebTarget;
+import jakarta.ws.rs.core.Response;
 import org.junit.Rule;
 import org.junit.Test;
-
-import javax.ws.rs.client.ClientBuilder;
-import javax.ws.rs.client.WebTarget;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -71,13 +69,13 @@ public class JerseyAppJunit4IT {
 
         // added as a part of a package
         Response r1 = base.path("/jt/jr/1").request().get();
-        assertEquals(Status.OK.getStatusCode(), r1.getStatus());
+        assertEquals(Response.Status.OK.getStatusCode(), r1.getStatus());
         String expected1 = "bootique-jersey-resource1:--config=src/test/resources/io/bootique/bom/jersey/test.yml,--server";
         assertEquals(expected1, r1.readEntity(String.class));
 
         // added as individual resource
         Response r2 = base.path("/jt/jr/2").request().get();
-        assertEquals(Status.OK.getStatusCode(), r2.getStatus());
+        assertEquals(Response.Status.OK.getStatusCode(), r2.getStatus());
         String expected2 = "bootique-jersey-resource2:--config=src/test/resources/io/bootique/bom/jersey/test.yml,--server";
         assertEquals(expected2, r2.readEntity(String.class));
     }