+++ bootique-bom-tests-all/src/test/java/io/bootique/bom/jetty/JettyAppJunit4IT.java
@@ -22,16 +22,13 @@ package io.bootique.bom.jetty;
 import io.bootique.jetty.JettyModule;
 import io.bootique.test.junit.BQTestFactory;
 import io.bootique.test.junit.TestIO;
+import jakarta.ws.rs.client.ClientBuilder;
+import jakarta.ws.rs.client.WebTarget;
+import jakarta.ws.rs.core.Response;
 import org.junit.Rule;
 import org.junit.Test;
 
-import javax.ws.rs.client.ClientBuilder;
-import javax.ws.rs.client.WebTarget;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 public class JettyAppJunit4IT {
 
@@ -62,15 +59,15 @@ public class JettyAppJunit4IT {
         WebTarget base = ClientBuilder.newClient().target("http://localhost:11234/");
 
         Response r1 = base.path("/testc").request().get();
-        assertEquals(Status.OK.getStatusCode(), r1.getStatus());
+        assertEquals(Response.Status.OK.getStatusCode(), r1.getStatus());
         String expected1 = String.format("bom_filter_before%nbom_servlet_query_string: null%nbom_filter_after%n");
         assertEquals(expected1, r1.readEntity(String.class));
 
         Response r2 = base.path("/testb").request().get();
-        assertEquals(Status.NOT_FOUND.getStatusCode(), r2.getStatus());
+        assertEquals(Response.Status.NOT_FOUND.getStatusCode(), r2.getStatus());
 
         Response r3 = base.path("/testc").queryParam("p", "v").request().get();
-        assertEquals(Status.OK.getStatusCode(), r3.getStatus());
+        assertEquals(Response.Status.OK.getStatusCode(), r3.getStatus());
         String expected3 = String.format("bom_filter_before%nbom_servlet_query_string: p=v%nbom_filter_after%n");
         assertEquals(expected3, r3.readEntity(String.class));
     }