+++ RELEASE-NOTES.md
@@ -1,3 +1,4 @@
 ## 3.0-RC1
 
 * #7 Missing "bootique-tapestry59-jakarta"
+* #8 Missing CXF Jakarta modules